fix bug about Assembly
