+++ src/dotnet-aries-doc/Program.cs
@@ -2,16 +2,11 @@
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.CommandLineUtils;
 using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.DependencyModel;
-using Microsoft.Extensions.DependencyModel.Resolution;
 using Newtonsoft.Json;
 using Pandv.AriesDoc.Generator;
 using System;
-using System.Collections.Generic;
 using System.IO;
 using System.Linq;
-using System.Reflection;
-using System.Runtime.Loader;
 
 namespace Pandv.AriesDoc
 {
@@ -109,13 +104,8 @@ namespace Pandv.AriesDoc
         public static void Generate(string dllDirectory, string docDirectory, string startupName,
             string baseUri, string version)
         {
-            var assemblies = Directory.EnumerateFiles(dllDirectory)
-                .Where(i => i.EndsWith(".dll"))
-                //.Select(i => AssemblyLoadContext.Default.LoadFromAssemblyPath(i))
-                .Select(i => new AssemblyResolver(i).Assembly)
-                .ToArray();
-            assemblies.ToList().ForEach(i => DependencyContext.Load(i));
-            var startupType = assemblies.SelectMany(i => i.ExportedTypes)
+            var startupType = new AssemblyResolver(dllDirectory).Assemblies
+                .SelectMany(i => i.ExportedTypes)
                 .FirstOrDefault(x => string.Equals(x.Name, startupName));
             var webHostBuilder = WebHost.CreateDefaultBuilder(new string[0])
                 .UseStartup(startupType)
@@ -130,68 +120,6 @@ namespace Pandv.AriesDoc
                 .UseUrls(baseUri)
                 .Build()
                 .GeneratorDoc(docDirectory, baseUri);
-        }
-    }
-
-    internal sealed class AssemblyResolver : IDisposable
-    {
-        private readonly ICompilationAssemblyResolver assemblyResolver;
-        private readonly DependencyContext dependencyContext;
-        private readonly AssemblyLoadContext loadContext;
-
-        public AssemblyResolver(string path)
-        {
-            this.Assembly = AssemblyLoadContext.Default.LoadFromAssemblyPath(path);
-            this.dependencyContext = DependencyContext.Load(this.Assembly);
-
-            this.assemblyResolver = new CompositeCompilationAssemblyResolver
-                                    (new ICompilationAssemblyResolver[]
-            {
-            new AppBaseCompilationAssemblyResolver(Path.GetDirectoryName(path)),
-            new ReferenceAssemblyPathResolver(),
-            new PackageCompilationAssemblyResolver()
-            });
-
-            this.loadContext = AssemblyLoadContext.GetLoadContext(this.Assembly);
-            this.loadContext.Resolving += OnResolving;
-        }
-
-        public Assembly Assembly { get; }
-
-        public void Dispose()
-        {
-            this.loadContext.Resolving -= this.OnResolving;
-        }
-
-        private Assembly OnResolving(AssemblyLoadContext context, AssemblyName name)
-        {
-            bool NamesMatch(RuntimeLibrary runtime)
-            {
-                return string.Equals(runtime.Name, name.Name, StringComparison.OrdinalIgnoreCase);
-            }
-
-            RuntimeLibrary library =
-                this.dependencyContext.RuntimeLibraries.FirstOrDefault(NamesMatch);
-            if (library != null)
-            {
-                var wrapper = new CompilationLibrary(
-                    library.Type,
-                    library.Name,
-                    library.Version,
-                    library.Hash,
-                    library.RuntimeAssemblyGroups.SelectMany(g => g.AssetPaths),
-                    library.Dependencies,
-                    library.Serviceable);
-
-                var assemblies = new List<string>();
-                this.assemblyResolver.TryResolveAssemblyPaths(wrapper, assemblies);
-                if (assemblies.Count > 0)
-                {
-                    return this.loadContext.LoadFromAssemblyPath(assemblies[0]);
-                }
-            }
-
-            return null;
         }
     }
 }
